+++ packages/solid/src/index.ts
@@ -1,0 +1,8 @@
+export { createClassed, classed } from "./classed";
+
+export type {
+  InferVariantProps,
+  ClassedComponentType,
+  VariantProps,
+  DerivedComponentType,
+} from "./types";